feat: relocate passport.use to src/passport
