+++ server/src/passport/passport.ts
@@ -1,0 +1,32 @@
+import passport from "passport";
+import crypto from "crypto";
+import {checkAuthData} from "../check/authData/authData";
+import JWTstrategy, {VerifiedCallback} from 'passport-jwt';
+import {ExtractJwt} from 'passport-jwt';
+import strategy from 'passport-local'
+let LocalStrategy = strategy.Strategy
+import config from 'config'
+
+passport.use('login', new LocalStrategy({
+        usernameField: 'email',
+        passwordField: 'password'
+    },
+    async function (username, password, done) {
+        let hashPass = crypto.createHash('sha256').update(password).digest('hex');
+        let authResult=await checkAuthData({email:username,password:hashPass})
+        return done(null, authResult);
+    }
+));
+
+
+passport.use( new JWTstrategy.Strategy({
+    secretOrKey: config.get('secretOrKey'),
+    jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken()
+}, async (token, done: VerifiedCallback) => {
+    console.log('token on use '+token.id)
+    try {
+        return done(null, token.id);
+    } catch (e) {
+        return done(e);
+    }
+}))
